user/requirements.txt
boto3==1.26.66
PyJWT==2.4.0
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.0.1